requirements.txt
llama-parse>=0.5.14
python-dotenv>=1.0.1 